REFACTOR: Extract generic types and constants from model graph vis.
