+++ packages/frontend/src/stdlib/analyses/model_graph.tsx
@@ -5,24 +5,24 @@ import type { ModelAnalysisProps } from "../../analysis";
 import type { ModelJudgment } from "../../model";
 import type { ModelAnalysisMeta, ModelTypeMeta, Theory } from "../../theory";
 import { DownloadSVGButton, GraphvizSVG, type SVGRefProp } from "../../visualization";
+import {
+    type GraphContent,
+    type GraphvizAttributes,
+    defaultEdgeAttributes,
+    defaultGraphAttributes,
+    defaultNodeAttributes,
+    graphvizEngine,
+} from "./graph";
 
 import textStyles from "../text_styles.module.css";
 import baseStyles from "./base_styles.module.css";
-
-/** Configuration for a graph visualization of a model. */
-export type ModelGraphContent = {
-    tag: "graph";
-
-    /** Layout engine for graph. */
-    layout: "graphviz-directed" | "graphviz-undirected";
-};
 
 /** Configure a graph visualization for use with models of a double theory. */
 export function configureModelGraph(options: {
     id: string;
     name: string;
     description?: string;
-}): ModelAnalysisMeta<ModelGraphContent> {
+}): ModelAnalysisMeta<GraphContent> {
     const { id, name, description } = options;
     return {
         id,
@@ -49,7 +49,7 @@ may be added in the future.
 export function ModelGraph(
     props: {
         title?: string;
-    } & ModelAnalysisProps<ModelGraphContent>,
+    } & ModelAnalysisProps<GraphContent>,
 ) {
     const [svgRef, setSvgRef] = createSignal<SVGSVGElement>();
 
@@ -80,16 +80,6 @@ export function ModelGraph(
             </Show>
         </div>
     );
-}
-
-export function graphvizEngine(layout: ModelGraphContent["layout"]) {
-    let engine!: Viz.RenderOptions["engine"];
-    if (layout === "graphviz-directed") {
-        engine = "dot";
-    } else if (layout === "graphviz-undirected") {
-        engine = "neato";
-    }
-    return engine;
 }
 
 /** Visualize a model of a double theory as a graph using Graphviz.
@@ -172,13 +162,6 @@ export function modelToGraphviz(
     };
 }
 
-/** Top-level attributes of a Graphviz graph. */
-export type GraphvizAttributes = {
-    graph?: Viz.Graph["graphAttributes"];
-    node?: Viz.Graph["nodeAttributes"];
-    edge?: Viz.Graph["edgeAttributes"];
-};
-
 const cssClass = (meta?: ModelTypeMeta): string =>
     [...(meta?.svgClasses ?? []), ...(meta?.textClasses ?? [])].join(" ");
 
@@ -186,20 +169,3 @@ const cssClass = (meta?: ModelTypeMeta): string =>
 // Graphviz a monospace font if and only if we're using one.
 const fontname = (meta?: ModelTypeMeta) =>
     meta?.textClasses?.includes(textStyles.code) ? "Courier" : "Helvetica";
-
-const defaultGraphAttributes = {
-    nodesep: "0.5",
-};
-
-const defaultNodeAttributes = {
-    // XXX: How to set the font size?
-    fontsize: "20",
-    shape: "box",
-    width: 0,
-    height: 0,
-};
-
-const defaultEdgeAttributes = {
-    fontsize: "20",
-    sep: "5",
-};
