+++ packages/frontend/src/stdlib/analyses/submodel_graphs.tsx
@@ -8,7 +8,8 @@ import type { ModelAnalysisProps } from "../../analysis";
 import { IconButton } from "../../components";
 import type { ModelJudgment } from "../../model";
 import type { ModelAnalysisMeta, Theory } from "../../theory";
-import { type GraphvizAttributes, ModelGraphviz } from "./model_graph";
+import type { GraphvizAttributes } from "./graph";
+import { ModelGraphviz } from "./model_graph";
 
 import baseStyles from "./base_styles.module.css";
 import "./submodel_graphs.css";
